requirements.txt
boto3
pytest
moto
ulid-py
python-multipart
fastapi
uvicorn